config.py
import os
from dotenv import load_dotenv

load_dotenv()  # reads your .env file and loads all variables

GROQ_API_KEY       = os.getenv("GROQ_API_KEY")
TWILIO_ACCOUNT_SID = os.getenv("TWILIO_ACCOUNT_SID")
TWILIO_AUTH_TOKEN  = os.getenv("TWILIO_AUTH_TOKEN")
DATABASE_URL       = os.getenv("DATABASE_URL", "postgresql://localhost/callcenter")
