requirements.txt
flask
groq
twilio
psycopg2-binary
chromadb
python-dotenv
sentence-transformers
